requirements.txt
python-twitter
wordcloud
numpy
image